Update the test modules and methods
