+++ tests/evidence_string_generation/test_clinvar.py
@@ -1,4 +1,3 @@
-from datetime import datetime
 import unittest
