+++ tests/evidence_string_generation/test_consequence_type.py
@@ -1,6 +1,5 @@
 import unittest
 
-import os
 from collections import defaultdict
 
 from eva_cttv_pipeline.evidence_string_generation import consequence_type as CT
